Create README.md
Design an API which Return data as Fizz Buzz Pattern: A program that prints each number from 1 to 100 on a new line. For each multiple of 3, print "Fizz" instead of the number. For each multiple of 5, print "Buzz" instead of the number. For numbers which are multiples of both 3 and 5, print "FizzBuzz" instead of the number. More on Fizz Buzz: https://en.wikipedia.org/wiki/Fizz_buzz BONUS: Validations on User parameters BONUS: If you can host the API somewhere to check the functionality BONUS: Unit Tests & Integration Tests